+++ src/components/sheets/form.tsx
@@ -12,7 +12,7 @@ import {
 import { Button } from '@/components/ui/button';
 import { Spinner } from '@/components/ui/spinner';
 
-function SheetWithForm({
+export default function SheetForm({
   sheetOpen,
   sheetClose,
   formId,
@@ -66,48 +66,3 @@ function SheetWithForm({
     </Sheet>
   );
 }
-
-function SheetWithContent({
-  sheetOpen,
-  sheetClose,
-  children,
-  title,
-  description,
-  size = 'default',
-}: {
-  sheetOpen: boolean;
-  sheetClose: () => void;
-  children: React.ReactNode;
-  title: string;
-  description?: string;
-  size?: 'default' | 'medium' | 'large' | 'xlarge';
-}) {
-  if (!sheetOpen) return null;
-
-  let sheetSizeClasses = 'sm:max-w-sm';
-
-  if (size === 'medium') sheetSizeClasses = 'sm:max-w-lg';
-  if (size === 'large') sheetSizeClasses = 'sm:max-w-2xl';
-  if (size === 'xlarge') sheetSizeClasses = 'sm:max-w-3xl';
-
-  return (
-    <Sheet open={sheetOpen} onOpenChange={(open) => !open && sheetClose()}>
-      <SheetContent className={sheetSizeClasses}>
-        <SheetHeader>
-          <SheetTitle>{title}</SheetTitle>
-          {description && <SheetDescription>{description}</SheetDescription>}
-        </SheetHeader>
-        <div className="flex-1 overflow-y-auto px-4">{children}</div>
-        <SheetFooter>
-          <SheetClose asChild>
-            <Button variant="outline" className="cursor-pointer">
-              Close
-            </Button>
-          </SheetClose>
-        </SheetFooter>
-      </SheetContent>
-    </Sheet>
-  );
-}
-
-export { SheetWithForm, SheetWithContent };